v0.4.4.0 Initial Release

+++ Installer/InstallerWindow.xaml.cs
@@ -21,19 +21,23 @@ namespace Installer
             InitializeComponent();
 
             string assemblyVersion = Assembly.GetExecutingAssembly().GetName().Version.ToString();
-            //assemblyVersion = assemblyVersion.Substring(0, assemblyVersion.LastIndexOf('.')) + "";
+ //           assemblyVersion = assemblyVersion.Substring(0, assemblyVersion.LastIndexOf('.')) + "";
             Title += " (" + assemblyVersion + ")";
 
-            if (Directory.Exists(Parameters.appDir))
+            messageQueue = new Queue<string>();
+            worker = new InstallerWorker(messageQueue);
+            if (worker.app2020Installed)
+            {
+                btnInstall.Content = "Upgrade!";
+                if (worker.appInstalled) chkMigrate2020Config.Visibility = Visibility.Visible;
+            }
+            else if (worker.appInstalled)
                 btnInstall.Content = "Update!";
             else
             {
                 btnRemove.IsEnabled = false;
                 btnRemove.Visibility = Visibility.Hidden;
             }
-
-            messageQueue = new Queue<string>();
-            worker = new InstallerWorker(messageQueue);
             timer = new DispatcherTimer
             {
                 Interval = TimeSpan.FromMilliseconds(200)
@@ -78,7 +82,19 @@ namespace Installer
                     MessageBox.Show($"Please stop {Parameters.appName} and try again.", $"{Parameters.appName} is running!", MessageBoxButton.OK, MessageBoxImage.Warning);
                     return;
                 }
-
+                if (InstallerFunctions.GetProcessRunning(Parameters.appName2020))
+                {
+                    MessageBox.Show($"Please stop {Parameters.appName2020} and try again.", $"{Parameters.appName2020} is running!", MessageBoxButton.OK, MessageBoxImage.Warning);
+                    return;
+                }
+                if (worker.app2020Installed)
+                {
+                    if (worker.appInstalled)
+                    {
+                        if (MessageBox.Show("Upgrading to MSFS_AutoFPS will uninstall MSFS2020_AutoFPS.\nYou have chosen to " + (worker.Migrate2020Config ? "transfer the config file from MSFS2020_AutoFPS to MSFS_AutoFPS to use as the MSFS 2020 config." : "use the MSFS 2020 config file already in MSFS_AutoFPS.") + " Are you sure?", "Existing MSFS2020_AutoFPS and MSFS_AutoFPS installations detected!", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.No) return;
+                    }
+                    else MessageBox.Show("Upgrading to MSFS_AutoFPS will uninstall MSFS2020_AutoFPS.\nYour existing config file from MSFS2020_AutoFPS will transfer to MSFS_AutoFPS to use as the MSFS 2020 config.", "Existing MSFS2020_AutoFPS installation detected!", MessageBoxButton.OK, MessageBoxImage.Information); 
+                }
                 chkDesktopLink_Click(null, null);
                 radio_Click(null, null);
 
@@ -108,9 +124,11 @@ namespace Installer
 
                 try
                 {
-                    Directory.Delete(Parameters.appDir, true);
+                    if (Directory.Exists(Parameters.binDir))
+                        Directory.Delete(Parameters.appDir, true);
                     InstallerFunctions.AutoStartExe(true);
-                    InstallerFunctions.AutoStartFsuipc(true);
+                    InstallerFunctions.AutoStartFsuipc(true, true);
+                    InstallerFunctions.RemoveDesktopLink(Parameters.appName);
                 }
                 catch (Exception ex)
                 {
@@ -138,6 +156,11 @@ namespace Installer
                 worker.CfgAutoStart = AutoStart.EXE;
             else if (radioRemove.IsChecked == true)
                 worker.CfgAutoStart = AutoStart.REMOVE;
+        }
+
+        private void chkMigrate2020Config_Click(object sender, RoutedEventArgs e)
+        {
+            worker.Migrate2020Config = chkMigrate2020Config.IsChecked ?? false;
         }
     }
 }
